requirements/base.txt
annotated-types==0.6.0
anyio==3.7.1
cffi==1.16.0
click==8.1.7
cryptography==41.0.5
ecdsa==0.18.0
fastapi==0.104.1
h11==0.14.0
idna==3.5
passlib==1.7.4
pyasn1==0.5.1
pycparser==2.21
pydantic_core==2.14.5
pydantic-settings==2.1.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
networkx==3.2.1
matplotlib==3.8.3
greenlet==3.0.3

# DB
alembic==1.12.0
asyncpg==0.28.0
psycopg==3.1.10
psycopg-binary==3.1.10
SQLAlchemy==2.0.23
